made rout put books based on roles with util handler images used in admin put book

+++ server/src/lib/cloud/delete.ts
@@ -22,6 +22,8 @@ export const delArrCloud = async (
   publicIDs: string[],
   resource?: ResourceType
 ) => {
+  if (!publicIDs?.length) return;
+
   try {
     const res = await cloud.api.delete_resources(publicIDs, {
       resource_type: resource ?? "image",
